requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-redis-cache==1.7.1
django-render-block==0.5
django-templated-email==2.2.0
djangorestframework==3.6.3
idna==2.5
kombu==4.1.0
psycopg2==2.7.2
pytz==2017.2
redis==2.10.5
requests==2.18.1
six==1.10.0
urllib3==1.21.1
vine==1.1.4
